@@ -7,12 +7,16 @@ use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        //$this->middleware(Authenticate::class);
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        echo 'UserController';
         $users = User::with('profile')->get();
         //dd($users);
         return view('users.users', compact('users'));
@@ -37,16 +41,16 @@ class UserController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(User $user)
+    public function show(string $id)
     {
         $user = User::with('profile')->findOrFail($id);
-        dd($user);
+        return view('users.profile', compact('user'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(User $user)
+    public function edit(string $id)
     {
         //
     }
@@ -54,15 +58,15 @@ class UserController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, string $id)
     {
-        //
+        $user = auth()->user();
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user)
+    public function destroy(string $id)
     {
         //
     }
