@@ -2,6 +2,7 @@
 
 namespace App\Providers;
 
+use Faker\Factory;
 use Illuminate\Database\Schema\Builder;
 use Illuminate\Support\ServiceProvider;
 
@@ -12,7 +13,9 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        //
+        $this->app->singleton(Generator::class, function () {
+            return Factory::create('pt_BR');
+        });
     }
 
     /**
