@@ -13,28 +13,27 @@ class ProfileFactory extends Factory
     {
 
         return [
-            'user_id' => $this->faker->unique()->numberBetween($min = 1, $max = 10),
+            'user_id' => $this->faker->unique()->numberBetween(1,11),
             'uuid' => $this->faker->uuid(),
-            'profile_image' =>  $this->faker->imageUrl($width = 50, $height = 50),
-            'profile_cpf' => $this->faker->cpf,
-            'profile_rg' => $this->faker->rg,
+            'profile_image' =>  $this->faker->imageUrl(150,150),
+            'profile_cpf' => $this->faker->cpf(),
+            'profile_rg' => $this->faker->rg(),
             'profile_rg_emit' => $this->faker->regexify('[A-Z]{2}'),
-            'profile_birthday' => $this->faker->date($format = 'Y-m-d', $max = '-18 years'),
+            'profile_birthday' => $this->faker->date('Y-m-d','-18 years'),
             'profile_email' => $this->faker->unique()->companyEmail, //,  $this->faker->unique()->safeEmail,
             'profile_address' => $this->faker->streetAddress,
             'profile_number' => $this->faker->buildingNumber,
             'profile_neighborhood' => $this->faker->citySuffix,
             'profile_city' => $this->faker->city,
-            'profile_estate' => $this->faker->countryCode(),
+            'profile_estate' => $this->faker->region(),
             'profile_country' => $this->faker->country,
             'profile_cep' => $this->faker->postcode,
-            'profile_telephone1' => $this->faker->unique()->phoneNumber,
-            'profile_telephone2' => $this->faker->unique()->phoneNumber,
-            'profile_telephone3' => $this->faker->unique()->phoneNumber,
-            'profile_telephone4' => $this->faker->unique()->phoneNumber,
-            'profile_obs' => $this->faker->text,
-            //'profile_file' => $this->faker->file(storage_path('app/public'), storage_path('app/public/uploads'), false),
-            'profile_file' => $this->faker->imageUrl($width = 400, $height = 400),
+            'profile_telephone1' => $this->faker->cellphoneNumber(),
+            'profile_telephone2' => $this->faker->phone(),
+            'profile_telephone3' => $this->faker->landlineNumber(false),
+            'profile_telephone4' => $this->faker->phoneNumberCleared(),
+            'profile_obs' => $this->faker->realText(),
+            'profile_file' => $this->faker->imageUrl(400, 400),
         ];
     }
 }
