@@ -2,6 +2,7 @@
 
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -16,7 +17,8 @@
             width: 100%;
         }
 
-        th, td {
+        th,
+        td {
             border: 1px solid #ddd;
             padding: 8px;
             text-align: left;
@@ -33,59 +35,67 @@
         }
     </style>
 </head>
+
 <body>
-<div class="container mt-4">
-    <h1>List of Users</h1>
+    <div class="container mt-4">
+        <h1>List of Users</h1>
 
-    @if ($users->isEmpty())
-        <p>No users available.</p>
-    @else
-        <table class="table">
-            <thead>
-            <tr>
-                <th>Profile Image</th>
-                <th>Name</th>
-                <th>Email</th>
-                <th>UUID</th>
-                <th>Address</th>
-                <th>Phone</th>
-            </tr>
-            </thead>
-            <tbody>
-            @foreach ($users as $user)
-                <tr>
-                    <td>
-                        @if ($user->profile && $user->profile->profile_image)
-                            <img src="{{ asset($user->profile->profile_image) }}" alt="Profile Image" class="profile-image">
-                        @else
-                            No image
-                        @endif
-                    </td>
-                    <td>{{ $user->name }}</td>
-                    <td>{{ $user->email }}</td>
-                    <td>{{ $user->uuid }}</td>
-                    <td>
-                        @if ($user->profile)
-                            {{ $user->profile->profile_address }}
-                        @else
-                            No profile available
-                        @endif
-                    </td>
-                    <td>
-                        @if ($user->profile)
-                            {{ $user->profile->profile_telephone1 }}
-                        @else
-                            No profile available
-                        @endif
-                    </td>
-                </tr>
-            @endforeach
-            </tbody>
-        </table>
-    @endif
-</div>
+        @if ($users->isEmpty())
+            <p>No users available.</p>
+        @else
+            <table class="table">
+                <thead>
+                    <tr>
+                        <th>Profile Image</th>
+                        <th>ID</th>
+                        <th>Prof.ID</th>
+                        <th>UUID</th>
+                        <th>Name</th>
+                        <th>Email</th>
+                        <th>Address</th>
+                        <th>Phone</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    @foreach ($users as $user)
+                        <tr>
+                            <td>
+                                @if ($user->profile && $user->profile->profile_image)
+                                    <img src="{{ asset($user->profile->profile_image) }}" alt="Profile Image"
+                                        class="profile-image">
+                                @else
+                                    No image
+                                @endif
+                            </td>
+                            <td>{{ $user->id }}</td>
+                            <td>{{ $user->profile->id }}</td>
+                            <td>{{ $user->uuid }}</td>
+                            <td>{{ $user->name }}</td>
+                            <td>{{ $user->email }}</td>
+                            <td>
+                                @if ($user->profile)
+                                    {{ $user->profile->profile_address }}
+                                @else
+                                    No profile available
+                                @endif
+                            </td>
+                            <td>
+                                @if ($user->profile)
+                                    {{ $user->profile->profile_telephone1 }}
+                                @else
+                                    No profile available
+                                @endif
+                            </td>
+                        </tr>
+                    @endforeach
+                </tbody>
+            </table>
+        @endif
+        {{ dd($user) }}
+    </div>
 
-<!-- Bootstrap JS and Popper.js via CDN (required for Bootstrap JavaScript components) -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"></script>
+    <!-- Bootstrap JS and Popper.js via CDN (required for Bootstrap JavaScript components) -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"></script>
 </body>
+
 </html>
